backend/requirements.txt
black==22.6.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.0.1
colorama==0.4.4
dnspython==2.2.1
feather-format==0.4.1
Flask==2.0.1
Flask-Cors==3.0.10
gazpacho==1.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.9.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.22.0
pandas==1.3.1
pathspec==0.9.0
platformdirs==2.5.2
pyarrow==5.0.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.3
